refactor(library): Position이 변함에 따라 연관된 코드 변경
