+++ src/main/kotlin/library/ForbiddenThree.kt
@@ -1,4 +1,4 @@
-package domain.rule
+package domain.library
 
 import domain.CoordinateState
 import domain.Position
@@ -6,10 +6,10 @@ import domain.Position
 object ForbiddenThree {
     fun isForbiddenThree(board: List<List<CoordinateState>>, position: Position): Boolean {
         var openThreeCount = 0
-        openThreeCount += findHorizontal(board, position.getX(), position.getY())
-        openThreeCount += findDiagonal1(board, position.getX(), position.getY())
-        openThreeCount += findVertical(board, position.getX(), position.getY())
-        openThreeCount += findDiagonal2(board, position.getX(), position.getY())
+        openThreeCount += findHorizontal(board, position.x, position.y)
+        openThreeCount += findDiagonal1(board, position.x, position.y)
+        openThreeCount += findVertical(board, position.x, position.y)
+        openThreeCount += findDiagonal2(board, position.x, position.y)
         return openThreeCount >= 2
     }
